use pytest conftest.py instead of common.py
